added comments, broke out components, and re-organized a few files
