+++ src/components/EmptyStepSelection.js
@@ -1,12 +1,14 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
+
+// components
 import Checkbox from './Checkbox';
 
 const EmptyStepSelection = ({ stepName, isSelected, onClick, isDisabled }) => {
   const label = `no ${stepName}`;
 
   return (
-    <>
+    <React.Fragment>
       <Checkbox
         id={stepName}
         checked={isSelected}
@@ -15,7 +17,7 @@ const EmptyStepSelection = ({ stepName, isSelected, onClick, isDisabled }) => {
         onClick={onClick}
       />
       <div className="spacer2" />
-    </>
+    </React.Fragment>
   );
 };
 
@@ -24,10 +26,13 @@ EmptyStepSelection.defaultProps = {
 };
 
 EmptyStepSelection.propTypes = {
+  // required
   stepName: PropTypes.string.isRequired,
   isSelected: PropTypes.bool.isRequired,
-  isDisabled: PropTypes.bool,
-  onClick: PropTypes.func.isRequired
+  onClick: PropTypes.func.isRequired,
+
+  // optional
+  isDisabled: PropTypes.bool
 };
 
 export default EmptyStepSelection;
